Move `typeOf` to `util/core`
