+++ lib/sinon.js
@@ -44,7 +44,7 @@ exposeCoreUtils(exports, require("./sinon/util/core/index"));
 
 // TODO move these two into util/core.
 exports.extend = deprecated.wrap(require("./sinon/extend"), deprecated.defaultMsg("extend"));
-exports.typeOf = deprecated.wrap(require("./sinon/typeOf"), deprecated.defaultMsg("typeOf"));
+exports.typeOf = deprecated.wrap(require("./sinon/util/core/typeOf"), deprecated.defaultMsg("typeOf"));
 
 exports.assert = require("./sinon/assert");
 exports.collection = require("./sinon/collection");
