+++ lib/sinon/match.js
@@ -13,7 +13,7 @@ var deepEqual = require("./util/core/deep-equal").use(match); // eslint-disable-
 var every = require("./util/core/every");
 var functionName = require("./util/core/function-name");
 var iterableToString = require("./util/core/iterable-to-string");
-var typeOf = require("./typeOf");
+var typeOf = require("./util/core/typeOf");
 var valueToString = require("./util/core/value-to-string");
 
 var indexOf = Array.prototype.indexOf;
