+++ lib/sinon/util/core/iterable-to-string.js
@@ -1,5 +1,5 @@
 "use strict";
-var typeOf = require("../../typeOf");
+var typeOf = require("./typeOf");
 
 module.exports = function iterableToString(obj) {
     var representation = "";
