+++ lib/sinon/util/core/typeOf.js
