+++ test/util/core/typeOf-test.js
@@ -1,7 +1,7 @@
 "use strict";
 
 var referee = require("referee");
-var sinonTypeOf = require("../lib/sinon/typeOf");
+var sinonTypeOf = require("../../../lib/sinon/util/core/typeOf");
 var assert = referee.assert;
 
 describe("typeOf", function () {
